feat(ffi): split packages into multiple units
