+++ squire_lib/src/ffi_player.rs
@@ -1,0 +1,1 @@
+
