+++ squire_lib/src/player_registry.rs
@@ -1,6 +1,4 @@
-use std::{
-    collections::{HashMap, HashSet},
-};
+use std::collections::{HashMap, HashSet};
 
 use serde::{Deserialize, Serialize};
