+++ squire_lib/tests/management.rs
@@ -1,5 +1,2 @@
-
 #[cfg(test)]
-mod tests {
-    
-}
+mod tests {}
