+++ squire_lib/tests/matches.rs
@@ -1,5 +1,2 @@
-
 #[cfg(test)]
-mod tests {
-    
-}
+mod tests {}
